Second trained model
Trained for 30 epochs, with a decreased learning rate by 0.5 after the 15th and the 25th epochs (see scheduler in the notebook). 
Reaches a slightly better accuracy (95.5%) and a more balanced result concerning sensitivity and specificity. 
However, as all the results are lower than those of the paper, the other model is preferred as it achieves a better sensitivity than the model from the original paper.